# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Free Admission Api",
        "description": "My Free Admission Api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/v1/counselor/referral/": {
            "get": {
                "tags": [
                    "Counselor Referral"
                ],
                "summary": "Retrieve a list of counselor referrals",
                "description": "Get List Counselor Referral",
                "operationId": "fba0483041cb17c9724d4f7ccb6dbe1d",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "example name"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "example email"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "example phone"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "example address"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "example role"
                                                    },
                                                    "created_by": {
                                                        "type": "string",
                                                        "example": "example created_by"
                                                    },
                                                    "updated_by": {
                                                        "type": "string",
                                                        "example": "example updated_by"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/counselor/referral/store": {
            "post": {
                "tags": [
                    "Counselor Referral"
                ],
                "summary": "Create a new counselor referral",
                "description": "Store a new Counselor Referral",
                "operationId": "9fea22c6daeb7ab9ae20f99137dd84d6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "Counselor",
                                            "Agent",
                                            "Referrer"
                                        ]
                                    },
                                    "created_by": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "example name",
                                    "email": "example email",
                                    "phone": "example phone",
                                    "address": "example address",
                                    "role": "Counselor",
                                    "created_by": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "phone"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "address"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "Counselor"
                                        },
                                        "created_by": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "updated_by": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/counselor/referral/update/{id}": {
            "put": {
                "tags": [
                    "Counselor Referral"
                ],
                "summary": "Update an existing Counselor Referral",
                "description": "Update Counselor Referral",
                "operationId": "737ce4c7861c1a1db23c40404b3dc80c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "content": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "example name",
                                    "email": "example email",
                                    "phone": "example phone",
                                    "address": "example address",
                                    "role": "example role",
                                    "created_by": "example created_by",
                                    "updated_by": "example updated_by"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "phone"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "address"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "role"
                                        },
                                        "created_by": {
                                            "type": "number",
                                            "example": "created_by"
                                        },
                                        "updated_by": {
                                            "type": "number",
                                            "example": "updated_by"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/counselor/referral/delete/{id}": {
            "delete": {
                "tags": [
                    "Counselor Referral"
                ],
                "summary": "Delete a Counselor Referral",
                "description": "Delete Counselor Referral",
                "operationId": "19ca3b8d92d1fc1af50347d99f0eda03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Counselor Referral  deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete todo success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Counselor Referral not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/counselor/referral/show/{id}": {
            "get": {
                "tags": [
                    "Counselor Referral"
                ],
                "summary": "Get single  Counselor Referral",
                "description": "Returns list of Counselor Referral",
                "operationId": "aa79606bb87f2bce626a0ab92557849c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/counselor/": {
            "get": {
                "tags": [
                    "Config Search"
                ],
                "summary": "list of counselor referrals role",
                "operationId": "1946a78aba9e101dc296806a21434081",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer Authentication - Enter token in format (Bearer <your-token>)"
            }
        }
    },
    "security": {
        "1": {
            "Bearer": []
        }
    }
}